requirements.txt
requests
psycopg2-binary
apache-airflow
